OCR_requirements.txt
# ============================================================================
# PDF OCR System - Python Dependencies
# ============================================================================

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# OCR Engines
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
PyMuPDF==1.23.8

# Optional OCR Engines (install based on your hardware)
# For GPU support, install PyTorch first:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121  # CUDA 12.1
# pip install torch torchvision  # For MPS (Apple Silicon)
#
# Then install OCR engines:
# easyocr==1.7.0  # GPU support
# paddlepaddle-gpu==2.5.1  # CUDA GPU
# paddlepaddle==2.5.1  # CPU only
# paddleocr==2.7.0.3

# Task Queue
celery==5.3.4
redis==5.0.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Storage
minio==7.2.0
boto3==1.34.0

# Monitoring & Logging
python-json-logger==2.0.7
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
